refactor: ensure tweet can only be deleted by owner

+++ internal/services/tweet.go
@@ -1,6 +1,7 @@
 package services
 
 import (
+	"fmt"
 	"github.com/ellypaws/go-chirp/internal/models"
 	"github.com/ellypaws/go-chirp/pkg/db"
 )
@@ -9,7 +10,15 @@ func CreateTweet(tweet models.Tweet) error {
 	return db.CreateTweet(tweet)
 }
 
-func DeleteTweet(tweetID int) error {
+func DeleteTweet(tweetID, userID int) error {
+	tweet, err := db.FetchTweet(tweetID)
+	if err != nil {
+		return err
+	}
+	if tweet.UserID != userID {
+		return fmt.Errorf("user %d is not the owner of tweet %d", userID, tweetID)
+	}
+
 	return db.DeleteTweet(tweetID)
 }
